break out bdd code into a module and make it a optional dependency.
