+++ aiger_analysis/__init__.py
@@ -2,3 +2,4 @@
 from aiger_analysis.analysis import is_equal, is_satisfiable, is_valid
 from aiger_analysis.cadet import eliminate, is_true_QBF
 from aiger_analysis.abc import simplify
+from aiger_analysis.bdd import from_bdd, to_bdd
